scripts/requirements.txt
feedparser>=6.0
pyyaml>=6.0
jinja2>=3.1
markdown>=3.5
python-dateutil>=2.8
anthropic>=0.40
python-frontmatter>=1.0
certifi>=2024.0
